Migrate DatePicker to shadcn behind SHADCN_MIGRATION flag
