+++ src/i18n/de.json
@@ -1692,6 +1692,11 @@
       "placeholder_question": "Frage in {{language}} übersetzen"
     }
   },
+  "date_picker": {
+    "today": "Heute",
+    "go_to_selected": "Zur Auswahl",
+    "open_calendar": "Kalender öffnen"
+  },
   "date_period_picker": {
     "end": "Ende",
     "quick_links": {
