+++ src/ui/DatePicker.stories.tsx
@@ -2,6 +2,7 @@ import type { Meta, StoryObj } from '@storybook/nextjs-vite';
 import { useState } from 'react';
 
 import { DatePicker } from './DatePicker';
+import { Label } from './Label';
 
 const meta: Meta<typeof DatePicker> = {
   title: 'Components/DatePicker',
@@ -19,11 +20,17 @@ type Story = StoryObj<typeof meta>;
 
 export const Default: Story = {
   args: {
+    id: 'date-picker-story',
     selected: new Date('2024-01-01'),
   },
   render: function RenderComponent(args) {
     const [selected, setSelected] = useState(args.selected);
 
-    return <DatePicker {...args} selected={selected} onChange={setSelected} />;
+    return (
+      <div>
+        <Label htmlFor={args.id}>Date</Label>
+        <DatePicker {...args} selected={selected} onChange={setSelected} />
+      </div>
+    );
   },
 };
